+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-## [0.10] - DEV
+## [0.9.2] - 2019-08-05
 ### Added
 - Add a shortcuts window.
 - Add an option to toggle the dark theme.